requirements.txt
psycopg2-binary>=2.9
requests>=2.28
feedparser>=6.0
python-dotenv>=1.0
google-api-python-client>=2.100
google-generativeai>=0.3.0
google-auth-oauthlib
google-auth-httplib2
langdetect
isodate
yt-dlp>=2024.4.9
faster-whisper
